apps/backend/pyproject.toml
[project]
name = "core-backend"
version = "0.1.0"
description = "Core backend with Keycloak auth"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "email-validator>=2.0.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "python-multipart>=0.0.9",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
    "aioboto3>=13.1.1",
    "motor>=3.3.2",
    "jinja2>=3.1.5",
]



[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.uv]
dev-dependencies = []

[tool.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
